UI system rework, github ignore settings
UI reworked to now use an InterfaceManager class that stores all labels,
buttons and Icons using string IDs. InterfaceManager will handle all
rendering of the UI, and keep all of the different elements stored in
Maps that you can retrieve with the string ID. This has removed the 19
UI variables in Game.cpp and replaced them with 1 manager. .gitignore
file finally updated to try and clean up the repo.